Moved soft dependencies to their compatible plugin modules
