+++ chatimage-paper/src/main/java/me/chickensaysbak/chatimage/paper/PluginPaper.java
@@ -5,8 +5,8 @@ import me.chickensaysbak.chatimage.core.adapters.CommandAdapter;
 import me.chickensaysbak.chatimage.core.adapters.PlayerAdapter;
 import me.chickensaysbak.chatimage.core.adapters.PluginAdapter;
 import me.chickensaysbak.chatimage.core.adapters.YamlAdapter;
-import me.chickensaysbak.chatimage.core.softdepend.DiscordSRVHandler;
-import me.chickensaysbak.chatimage.core.softdepend.EssXDiscordHandler;
+import me.chickensaysbak.chatimage.paper.softdepend.DiscordSRVHandler;
+import me.chickensaysbak.chatimage.paper.softdepend.EssXDiscordHandler;
 import me.chickensaysbak.chatimage.paper.softdepend.PAPIHandler;
 import org.bstats.bukkit.Metrics;
 import org.bstats.charts.DrilldownPie;
