+++ chatimage-spigot/src/main/java/me/chickensaysbak/chatimage/spigot/PluginSpigot.java
@@ -7,8 +7,8 @@ import me.chickensaysbak.chatimage.core.adapters.CommandAdapter;
 import me.chickensaysbak.chatimage.core.adapters.PlayerAdapter;
 import me.chickensaysbak.chatimage.core.adapters.PluginAdapter;
 import me.chickensaysbak.chatimage.core.adapters.YamlAdapter;
-import me.chickensaysbak.chatimage.core.softdepend.DiscordSRVHandler;
-import me.chickensaysbak.chatimage.core.softdepend.EssXDiscordHandler;
+import me.chickensaysbak.chatimage.spigot.softdepend.DiscordSRVHandler;
+import me.chickensaysbak.chatimage.spigot.softdepend.EssXDiscordHandler;
 import me.chickensaysbak.chatimage.spigot.softdepend.PAPIHandler;
 import org.bstats.bukkit.Metrics;
 import org.bstats.charts.DrilldownPie;
