+++ chatimage-spigot/src/main/java/me/chickensaysbak/chatimage/spigot/softdepend/DiscordSRVHandler.java
@@ -1,6 +1,6 @@
 // ChatImage © 2023 ChickenSaysBak
 // This code is licensed under MIT license (see LICENSE file for details).
-package me.chickensaysbak.chatimage.core.softdepend;
+package me.chickensaysbak.chatimage.spigot.softdepend;
 
 import github.scarsz.discordsrv.DiscordSRV;
 import github.scarsz.discordsrv.api.Subscribe;
